requirements.txt
discord-py-interactions==5.0.1
requests==2.28.2
git+https://github.com/openai/whisper.git
python-dotenv==1.0.0
